Publish initial version
Includes two application modes, the server one which interacts with Twilio through
websockets and webhooks, and, the local mode which interacts with the computer's
microphone and speaker.

The server mode includes a simple page to trigger calls.

The input audio is submitted to Gemini Live API and the output is either played
through the Twilio connection or through the speaker.

There are a few demo prompts for testing purposes.

The call/process can be terminated from Gemini which detects words like "bye".

This has been tested in Spanish and English.

Adding more stages to the audio processing pipeline should be relatively simple.